+++ src/features/table/components/Toolbar/index.ts
@@ -3,6 +3,5 @@ export * from './RowsSelect';
 export * from './GlobalSearch';
 export * from './ToolbarTitle';
 export * from './ActionButtons';
-export * from './ResultsInfo';
 export * from './FilterBadges';
 export * from './SettingsPanel';